+++ packages/astro/src/assets/utils/transformToPath.ts
@@ -1,19 +1,18 @@
-import { basename, extname } from 'node:path';
+import { basename, dirname, extname } from 'node:path';
 import { deterministicString } from 'deterministic-object-hash';
 import { removeQueryString } from '../../core/path.js';
 import { shorthash } from '../../runtime/server/shorthash.js';
 import type { ImageTransform } from '../types.js';
 import { isESMImportedImage } from './imageKind.js';
 
-export function propsToFilename(transform: ImageTransform, hash: string) {
-	let filename = removeQueryString(
-		isESMImportedImage(transform.src) ? transform.src.src : transform.src
-	);
+export function propsToFilename(filePath: string, transform: ImageTransform, hash: string) {
+	let filename = decodeURIComponent(removeQueryString(filePath));
 	const ext = extname(filename);
-	filename = decodeURIComponent(basename(filename, ext));
+	filename = basename(filename, ext);
+	const prefixDirname = isESMImportedImage(transform.src) ? dirname(filePath) : '';
 
 	let outputExt = transform.format ? `.${transform.format}` : ext;
-	return `/${filename}_${hash}${outputExt}`;
+	return decodeURIComponent(`${prefixDirname}/${filename}_${hash}${outputExt}`);
 }
 
 export function hashTransform(