+++ packages/astro/src/assets/vite-plugin-assets.ts
@@ -9,6 +9,7 @@ import {
 	appendForwardSlash,
 	joinPaths,
 	prependForwardSlash,
+	removeBase,
 	removeQueryString,
 } from '../core/path.js';
 import { isServerLikeOutput } from '../prerender/utils.js';
@@ -68,14 +69,20 @@ export default function assets({
 					export { default as Picture } from "astro/components/Picture.astro";
 
 					export const imageConfig = ${JSON.stringify(settings.config.image)};
-					export const outDir = new URL(${JSON.stringify(
+					// This is used by the @astrojs/node integration to locate images.
+					// It's unused on other platforms, but on some platforms like Netlify (and presumably also Vercel)
+					// new URL("dist/...") is interpreted by the bundler as a signal to include that directory
+					// in the Lambda bundle, which would bloat the bundle with images.
+					// To prevent this, we mark the URL construction as pure,
+					// so that it's tree-shaken away for all platforms that don't need it.
+					export const outDir = /* #__PURE__ */ new URL(${JSON.stringify(
 						new URL(
 							isServerLikeOutput(settings.config)
 								? settings.config.build.client
 								: settings.config.outDir
 						)
 					)});
-					export const assetsDir = new URL(${JSON.stringify(settings.config.build.assets)}, outDir);
+					export const assetsDir = /* #__PURE__ */ new URL(${JSON.stringify(settings.config.build.assets)}, outDir);
 					export const getImage = async (options) => await getImageInternal(options, imageConfig);
 				`;
 				}
@@ -85,7 +92,7 @@ export default function assets({
 					return;
 				}
 
-				globalThis.astroAsset.addStaticImage = (options, hashProperties, originalPath) => {
+				globalThis.astroAsset.addStaticImage = (options, hashProperties, originalFSPath) => {
 					if (!globalThis.astroAsset.staticImages) {
 						globalThis.astroAsset.staticImages = new Map<
 							string,
@@ -96,13 +103,18 @@ export default function assets({
 						>();
 					}
 
-					// Rollup will copy the file to the output directory, this refer to this final path, not to the original path
+					// Rollup will copy the file to the output directory, as such this is the path in the output directory, including the asset prefix / base
 					const ESMImportedImageSrc = isESMImportedImage(options.src)
 						? options.src.src
 						: options.src;
 					const fileExtension = extname(ESMImportedImageSrc);
-					const pf = getAssetsPrefix(fileExtension, settings.config.build.assetsPrefix);
-					const finalOriginalImagePath = ESMImportedImageSrc.replace(pf, '');
+					const assetPrefix = getAssetsPrefix(fileExtension, settings.config.build.assetsPrefix);
+
+					// This is the path to the original image, from the dist root, without the base or the asset prefix (e.g. /_astro/image.hash.png)
+					const finalOriginalPath = removeBase(
+						removeBase(ESMImportedImageSrc, settings.config.base),
+						assetPrefix
+					);
 
 					const hash = hashTransform(
 						options,
@@ -111,21 +123,26 @@ export default function assets({
 					);
 
 					let finalFilePath: string;
-					let transformsForPath = globalThis.astroAsset.staticImages.get(finalOriginalImagePath);
+					let transformsForPath = globalThis.astroAsset.staticImages.get(finalOriginalPath);
 					let transformForHash = transformsForPath?.transforms.get(hash);
+
+					// If the same image has already been transformed with the same options, we'll reuse the final path
 					if (transformsForPath && transformForHash) {
 						finalFilePath = transformForHash.finalPath;
 					} else {
 						finalFilePath = prependForwardSlash(
-							joinPaths(settings.config.build.assets, propsToFilename(options, hash))
+							joinPaths(
+								isESMImportedImage(options.src) ? '' : settings.config.build.assets,
+								prependForwardSlash(propsToFilename(finalOriginalPath, options, hash))
+							)
 						);
 
 						if (!transformsForPath) {
-							globalThis.astroAsset.staticImages.set(finalOriginalImagePath, {
-								originalSrcPath: originalPath,
+							globalThis.astroAsset.staticImages.set(finalOriginalPath, {
+								originalSrcPath: originalFSPath,
 								transforms: new Map(),
 							});
-							transformsForPath = globalThis.astroAsset.staticImages.get(finalOriginalImagePath)!;
+							transformsForPath = globalThis.astroAsset.staticImages.get(finalOriginalPath)!;
 						}
 
 						transformsForPath.transforms.set(hash, {
@@ -137,7 +154,7 @@ export default function assets({
 					// The paths here are used for URLs, so we need to make sure they have the proper format for an URL
 					// (leading slash, prefixed with the base / assets prefix, encoded, etc)
 					if (settings.config.build.assetsPrefix) {
-						return encodeURI(joinPaths(pf, finalFilePath));
+						return encodeURI(joinPaths(assetPrefix, finalFilePath));
 					} else {
 						return encodeURI(prependForwardSlash(joinPaths(settings.config.base, finalFilePath)));
 					}