+++ packages/astro/src/content/index.ts
@@ -2,7 +2,12 @@ export { CONTENT_FLAG, PROPAGATED_ASSET_FLAG } from './consts.js';
 export { errorMap } from './error-map.js';
 export { attachContentServerListeners } from './server-listeners.js';
 export { createContentTypesGenerator } from './types-generator.js';
-export { contentObservable, getContentPaths, getDotAstroTypeReference } from './utils.js';
+export {
+	contentObservable,
+	getContentPaths,
+	getDotAstroTypeReference,
+	hasAssetPropagationFlag,
+} from './utils.js';
 export { astroContentAssetPropagationPlugin } from './vite-plugin-content-assets.js';
 export { astroContentImportPlugin } from './vite-plugin-content-imports.js';
 export { astroContentVirtualModPlugin } from './vite-plugin-content-virtual-mod.js';