+++ packages/db/src/core/integration/vite-plugin-db.ts
@@ -118,7 +118,8 @@ import { asDrizzleTable, createLocalDatabaseClient } from ${RUNTIME_IMPORT};
 ${shouldSeed ? `import { seedLocal } from ${RUNTIME_IMPORT};` : ''}
 ${shouldSeed ? integrationSeedImportStatements.join('\n') : ''}
 
-const dbUrl = ${JSON.stringify(dbUrl)};
+const dbUrl = import.meta.env.ASTRO_DATABASE_FILE ?? ${JSON.stringify(dbUrl)};
+
 export const db = createLocalDatabaseClient({ dbUrl });
 
 ${
@@ -134,7 +135,7 @@ ${
 
 export * from ${RUNTIME_CONFIG_IMPORT};
 
-${getStringifiedCollectionExports(tables)}`;
+${getStringifiedTableExports(tables)}`;
 }
 
 export function getStudioVirtualModContents({
@@ -168,16 +169,16 @@ export const db = await createRemoteDatabaseClient(${appTokenArg()}, ${dbUrlArg(
 
 export * from ${RUNTIME_CONFIG_IMPORT};
 
-${getStringifiedCollectionExports(tables)}
+${getStringifiedTableExports(tables)}
 	`;
 }
 
-function getStringifiedCollectionExports(tables: DBTables) {
+function getStringifiedTableExports(tables: DBTables) {
 	return Object.entries(tables)
 		.map(
-			([name, collection]) =>
+			([name, table]) =>
 				`export const ${name} = asDrizzleTable(${JSON.stringify(name)}, ${JSON.stringify(
-					collection
+					table
 				)}, false)`
 		)
 		.join('\n');