+++ packages/db/src/core/types.ts
@@ -12,13 +12,13 @@ import type {
 	numberColumnOptsSchema,
 	numberColumnSchema,
 	referenceableColumnSchema,
+	resolvedIndexSchema,
 	tableSchema,
 	textColumnOptsSchema,
 	textColumnSchema,
 } from './schemas.js';
 
-export type Indexes = Record<string, z.infer<typeof indexSchema>>;
-
+export type ResolvedIndexes = z.output<typeof dbConfigSchema>['tables'][string]['indexes'];
 export type BooleanColumn = z.infer<typeof booleanColumnSchema>;
 export type BooleanColumnInput = z.input<typeof booleanColumnSchema>;
 export type NumberColumn = z.infer<typeof numberColumnSchema>;
@@ -47,8 +47,10 @@ export type DBColumnInput =
 export type DBColumns = z.infer<typeof columnsSchema>;
 export type DBTable = z.infer<typeof tableSchema>;
 export type DBTables = Record<string, DBTable>;
+export type ResolvedDBTables = z.output<typeof dbConfigSchema>['tables'];
+export type ResolvedDBTable = z.output<typeof dbConfigSchema>['tables'][string];
 export type DBSnapshot = {
-	schema: Record<string, DBTable>;
+	schema: Record<string, ResolvedDBTable>;
 	version: string;
 };
 
@@ -67,7 +69,7 @@ export interface TableConfig<TColumns extends ColumnsConfig = ColumnsConfig>
 		columns: MaybeArray<Extract<keyof TColumns, string>>;
 		references: () => MaybeArray<z.input<typeof referenceableColumnSchema>>;
 	}>;
-	indexes?: Record<string, IndexConfig<TColumns>>;
+	indexes?: Array<IndexConfig<TColumns>> | Record<string, LegacyIndexConfig<TColumns>>;
 	deprecated?: boolean;
 }
 
@@ -75,9 +77,11 @@ interface IndexConfig<TColumns extends ColumnsConfig> extends z.input<typeof ind
 	on: MaybeArray<Extract<keyof TColumns, string>>;
 }
 
-/** @deprecated Use `TableConfig` instead */
-export type ResolvedCollectionConfig<TColumns extends ColumnsConfig = ColumnsConfig> =
-	TableConfig<TColumns>;
+/** @deprecated */
+interface LegacyIndexConfig<TColumns extends ColumnsConfig>
+	extends z.input<typeof resolvedIndexSchema> {
+	on: MaybeArray<Extract<keyof TColumns, string>>;
+}
 
 // We cannot use `Omit<NumberColumn | TextColumn, 'type'>`,
 // since Omit collapses our union type on primary key.