+++ packages/db/test/fixtures/local-prod/db/seed.ts
@@ -1,0 +1,8 @@
+import { asDrizzleTable } from '@astrojs/db/utils';
+import { User, db } from 'astro:db';
+
+export default async function () {
+	await db.batch([
+		db.insert(User).values([{ id: 'mario', username: 'Mario', password: 'itsame' }]),
+	]);
+}